UX and Accessibility
This is a rough draft of the UX and Accessibility lesson. I plan on doing another pass, but was going to do some
accessibility testing on the page prior to that and use any issues I found as examples in the final draft.

I also introduced a `.editorconfig` that unfortunately reformatted the whole codebase, so the diff is rather large. Not
ideal, but I'd rather not tweak values forever trying to get it right so going to accept the large diff.